git-svn-id: https://randi2.svn.sourceforge.net/svnroot/randi2/trunk@2553 ccfe945e-4522-0410-b90e-900a8cb55c58

+++ src/main/java/de/randi2/model/Login.java
@@ -37,7 +37,7 @@ public class Login extends AbstractDomainObject implements UserDetails {
 	
 	private final static long serialVersionUID = -6809229052570773439L;
 
-	public final static short MAX_WRONG_LOGINS=3;
+	public final static byte MAX_WRONG_LOGINS=3;
 	//15min
 	public final static int MILIS_TO_LOCK_USER = 900000;
